Move flux communication actions to ParallelAlgorithms
