+++ tests/Unit/NumericalAlgorithms/DiscontinuousGalerkin/Actions/CMakeLists.txt
@@ -6,8 +6,6 @@ set(LIBRARY_SOURCES
   Actions/Test_ApplyBoundaryFluxesLocalTimeStepping.cpp
   Actions/Test_ApplyFluxes.cpp
   Actions/Test_ComputeNonconservativeBoundaryFluxes.cpp
-  Actions/Test_FluxCommunication.cpp
-  Actions/Test_FluxCommunicationLts.cpp
   Actions/Test_ImposeBoundaryConditions.cpp
   PARENT_SCOPE
   )
