+++ tests/Unit/ParallelAlgorithms/DiscontinuousGalerkin/CMakeLists.txt
@@ -4,6 +4,8 @@
 set(LIBRARY "Test_ParallelDiscontinuousGalerkin")
 
 set(LIBRARY_SOURCES
+  Test_FluxCommunication.cpp
+  Test_FluxCommunicationLts.cpp
   Test_HasReceivedFromAllMortars.cpp
   Test_InitializeDomain.cpp
   Test_InitializeInterfaces.cpp
